Vil at tester skal fungere etter at vi legger til ny G for mai 2023. Vil trekke ut mockTestMedGrunnbeløpFra2022 så den kan brukes i flere tester.
