+++ src/test/kotlin/no/nav/familie/ef/sak/beregning/ValiderOmregningServiceTest.kt
@@ -14,6 +14,7 @@ import no.nav.familie.ef.sak.repository.behandling
 import no.nav.familie.ef.sak.repository.fagsak
 import no.nav.familie.ef.sak.repository.saksbehandling
 import no.nav.familie.ef.sak.repository.vedtak
+import no.nav.familie.ef.sak.testutil.mockTestMedGrunnbeløpFra2022
 import no.nav.familie.ef.sak.tilkjentytelse.TilkjentYtelseRepository
 import no.nav.familie.ef.sak.vedtak.VedtakService
 import no.nav.familie.ef.sak.vedtak.domain.AktivitetType
@@ -53,9 +54,11 @@ class ValiderOmregningServiceTest {
         val saksbehandling = saksbehandling(fagsak = fagsak(), behandling(forrigeBehandlingId = UUID.randomUUID()))
         mockVedtakOgForrigeTilkjentYtelse(saksbehandling)
         mockNyTilkjentYtelse(saksbehandling, medRiktigBeløp = false)
+        mockTestMedGrunnbeløpFra2022 {
+            assertThatThrownBy { validerOmregningService.validerHarGammelGOgKanLagres(saksbehandling) }
+                .isInstanceOf(ApiFeil::class.java)
+        }
 
-        assertThatThrownBy { validerOmregningService.validerHarGammelGOgKanLagres(saksbehandling) }
-            .isInstanceOf(ApiFeil::class.java)
     }
 
     @Test
@@ -73,17 +76,17 @@ class ValiderOmregningServiceTest {
         mockVedtakOgForrigeTilkjentYtelse(saksbehandling)
 
         every { tilkjentYtelseRepository.findByBehandlingId(saksbehandling.id) } returns
-            lagTilkjentYtelse(
-                andelerTilkjentYtelse = listOf(
-                    lagAndelTilkjentYtelse(
-                        fraOgMed = LocalDate.of(2022, 4, 1),
-                        tilOgMed = LocalDate.of(2022, 8, 30),
-                        samordningsfradrag = 5000,
-                        beløp = 0,
+                lagTilkjentYtelse(
+                    andelerTilkjentYtelse = listOf(
+                        lagAndelTilkjentYtelse(
+                            fraOgMed = LocalDate.of(2022, 4, 1),
+                            tilOgMed = LocalDate.of(2022, 8, 30),
+                            samordningsfradrag = 5000,
+                            beløp = 0,
+                        ),
                     ),
-                ),
-                grunnbeløpsmåned = YearMonth.of(2021, 5),
-            )
+                    grunnbeløpsmåned = YearMonth.of(2021, 5),
+                )
 
         validerOmregningService.validerHarGammelGOgKanLagres(saksbehandling)
     }
@@ -101,7 +104,8 @@ class ValiderOmregningServiceTest {
 
         @Test
         internal fun `skal ikke validere hvis det er maskinell g-omregning`() {
-            val saksbehandling = saksbehandling(årsak = BehandlingÅrsak.G_OMREGNING).copy(opprettetAv = SYSTEM_FORKORTELSE)
+            val saksbehandling =
+                saksbehandling(årsak = BehandlingÅrsak.G_OMREGNING).copy(opprettetAv = SYSTEM_FORKORTELSE)
             val vedtak = InnvilgelseOvergangsstønad(null, null, listOf())
 
             validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling)
@@ -109,7 +113,8 @@ class ValiderOmregningServiceTest {
 
         @Test
         internal fun `skal ikke validere hvis årsak er nye opplysninger`() {
-            val saksbehandling = saksbehandling(årsak = BehandlingÅrsak.NYE_OPPLYSNINGER).copy(opprettetAv = "saksbehandler")
+            val saksbehandling =
+                saksbehandling(årsak = BehandlingÅrsak.NYE_OPPLYSNINGER).copy(opprettetAv = "saksbehandler")
             val vedtak = InnvilgelseOvergangsstønad(null, null, listOf())
 
             validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling)
@@ -123,7 +128,7 @@ class ValiderOmregningServiceTest {
             val aktivitet = AktivitetType.BARNET_ER_SYKT
             val periodeType = VedtaksperiodeType.PERIODE_FØR_FØDSEL
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
-                innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
+                    innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
 
             val vedtak = innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
 
@@ -139,15 +144,15 @@ class ValiderOmregningServiceTest {
             val periodeType = VedtaksperiodeType.PERIODE_FØR_FØDSEL
 
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
-                innvilge(
-                    VedtaksperiodeDto(
-                        fra,
-                        til,
-                        Månedsperiode(fra, til),
-                        AktivitetType.MIGRERING,
-                        VedtaksperiodeType.MIGRERING,
-                    ),
-                )
+                    innvilge(
+                        VedtaksperiodeDto(
+                            fra,
+                            til,
+                            Månedsperiode(fra, til),
+                            AktivitetType.MIGRERING,
+                            VedtaksperiodeType.MIGRERING,
+                        ),
+                    )
 
             val vedtak = innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
             validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling)
@@ -160,7 +165,12 @@ class ValiderOmregningServiceTest {
 
             val vedtak = InnvilgelseOvergangsstønad(null, null, listOf())
 
-            assertThatThrownBy { validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling) }
+            assertThatThrownBy {
+                validerOmregningService.validerHarSammePerioderSomTidligereVedtak(
+                    vedtak,
+                    saksbehandling
+                )
+            }
                 .isInstanceOf(ApiFeil::class.java)
                 .hasMessageContaining("ikke har noen tidligere perioder")
         }
@@ -173,19 +183,24 @@ class ValiderOmregningServiceTest {
             val aktivitet = AktivitetType.BARNET_ER_SYKT
             val periodeType = VedtaksperiodeType.PERIODE_FØR_FØDSEL
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
-                innvilge(
-                    VedtaksperiodeDto(
-                        fra,
-                        til,
-                        Månedsperiode(fra, til),
-                        AktivitetType.IKKE_AKTIVITETSPLIKT,
-                        VedtaksperiodeType.MIGRERING,
-                    ),
-                )
+                    innvilge(
+                        VedtaksperiodeDto(
+                            fra,
+                            til,
+                            Månedsperiode(fra, til),
+                            AktivitetType.IKKE_AKTIVITETSPLIKT,
+                            VedtaksperiodeType.MIGRERING,
+                        ),
+                    )
 
             val vedtak = innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
 
-            assertThatThrownBy { validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling) }
+            assertThatThrownBy {
+                validerOmregningService.validerHarSammePerioderSomTidligereVedtak(
+                    vedtak,
+                    saksbehandling
+                )
+            }
                 .isInstanceOf(ApiFeil::class.java)
                 .hasMessageContaining("har annen aktivitet")
         }
@@ -198,19 +213,24 @@ class ValiderOmregningServiceTest {
             val aktivitet = AktivitetType.BARNET_ER_SYKT
             val periodeType = VedtaksperiodeType.PERIODE_FØR_FØDSEL
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
-                innvilge(
-                    VedtaksperiodeDto(
-                        fra,
-                        til,
-                        Månedsperiode(fra, til),
-                        AktivitetType.MIGRERING,
-                        VedtaksperiodeType.HOVEDPERIODE,
-                    ),
-                )
+                    innvilge(
+                        VedtaksperiodeDto(
+                            fra,
+                            til,
+                            Månedsperiode(fra, til),
+                            AktivitetType.MIGRERING,
+                            VedtaksperiodeType.HOVEDPERIODE,
+                        ),
+                    )
 
             val vedtak = innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
 
-            assertThatThrownBy { validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling) }
+            assertThatThrownBy {
+                validerOmregningService.validerHarSammePerioderSomTidligereVedtak(
+                    vedtak,
+                    saksbehandling
+                )
+            }
                 .isInstanceOf(ApiFeil::class.java)
                 .hasMessageContaining("annen type periode")
         }
@@ -224,20 +244,33 @@ class ValiderOmregningServiceTest {
             val aktivitet = AktivitetType.BARNET_ER_SYKT
             val periodeType = VedtaksperiodeType.PERIODE_FØR_FØDSEL
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
+                    innvilge(
+                        VedtaksperiodeDto(
+                            fra,
+                            til,
+                            Månedsperiode(fra, til),
+                            AktivitetType.MIGRERING,
+                            VedtaksperiodeType.HOVEDPERIODE,
+                        ),
+                    )
+
+            val vedtak =
                 innvilge(
                     VedtaksperiodeDto(
                         fra,
-                        til,
-                        Månedsperiode(fra, til),
-                        AktivitetType.MIGRERING,
-                        VedtaksperiodeType.HOVEDPERIODE,
-                    ),
+                        revurderingTil,
+                        Månedsperiode(fra, revurderingTil),
+                        aktivitet,
+                        periodeType
+                    )
                 )
 
-            val vedtak =
-                innvilge(VedtaksperiodeDto(fra, revurderingTil, Månedsperiode(fra, revurderingTil), aktivitet, periodeType))
-
-            assertThatThrownBy { validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling) }
+            assertThatThrownBy {
+                validerOmregningService.validerHarSammePerioderSomTidligereVedtak(
+                    vedtak,
+                    saksbehandling
+                )
+            }
                 .isInstanceOf(ApiFeil::class.java)
                 .hasMessageContaining("har annet tom-dato")
         }
@@ -248,19 +281,24 @@ class ValiderOmregningServiceTest {
             val fra = YearMonth.of(år, 5)
             val til = fra
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
-                innvilge(
-                    VedtaksperiodeDto(
-                        fra,
-                        til,
-                        Månedsperiode(fra, til),
-                        AktivitetType.MIGRERING,
-                        VedtaksperiodeType.HOVEDPERIODE,
-                    ),
-                )
+                    innvilge(
+                        VedtaksperiodeDto(
+                            fra,
+                            til,
+                            Månedsperiode(fra, til),
+                            AktivitetType.MIGRERING,
+                            VedtaksperiodeType.HOVEDPERIODE,
+                        ),
+                    )
 
             val vedtak = InnvilgelseOvergangsstønad(null, null, listOf())
 
-            assertThatThrownBy { validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling) }
+            assertThatThrownBy {
+                validerOmregningService.validerHarSammePerioderSomTidligereVedtak(
+                    vedtak,
+                    saksbehandling
+                )
+            }
                 .isInstanceOf(ApiFeil::class.java)
                 .hasMessageContaining("Antall vedtaksperioder er ulikt fra tidligere vedtak")
         }
@@ -273,7 +311,7 @@ class ValiderOmregningServiceTest {
             val aktivitet = AktivitetType.BARNET_ER_SYKT
             val periodeType = VedtaksperiodeType.PERIODE_FØR_FØDSEL
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
-                innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
+                    innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
 
             val vedtak = innvilge(VedtaksperiodeDto(fra, til, Månedsperiode(fra, til), aktivitet, periodeType))
 
@@ -289,19 +327,33 @@ class ValiderOmregningServiceTest {
             val aktivitet = AktivitetType.BARNET_ER_SYKT
             val periodeType = VedtaksperiodeType.PERIODE_FØR_FØDSEL
             every { vedtakHistorikkService.hentVedtakForOvergangsstønadFraDato(any(), any()) } returns
-                innvilge(
-                    VedtaksperiodeDto(
-                        fra,
-                        fra,
-                        Månedsperiode(fra, fra),
-                        AktivitetType.IKKE_AKTIVITETSPLIKT,
-                        VedtaksperiodeType.SANKSJON,
-                    ),
-                    VedtaksperiodeDto(andrePeriodeFra, til, Månedsperiode(andrePeriodeFra, til), aktivitet, periodeType),
-                )
+                    innvilge(
+                        VedtaksperiodeDto(
+                            fra,
+                            fra,
+                            Månedsperiode(fra, fra),
+                            AktivitetType.IKKE_AKTIVITETSPLIKT,
+                            VedtaksperiodeType.SANKSJON,
+                        ),
+                        VedtaksperiodeDto(
+                            andrePeriodeFra,
+                            til,
+                            Månedsperiode(andrePeriodeFra, til),
+                            aktivitet,
+                            periodeType
+                        ),
+                    )
 
             val vedtak =
-                innvilge(VedtaksperiodeDto(andrePeriodeFra, til, Månedsperiode(andrePeriodeFra, til), aktivitet, periodeType))
+                innvilge(
+                    VedtaksperiodeDto(
+                        andrePeriodeFra,
+                        til,
+                        Månedsperiode(andrePeriodeFra, til),
+                        aktivitet,
+                        periodeType
+                    )
+                )
 
             validerOmregningService.validerHarSammePerioderSomTidligereVedtak(vedtak, saksbehandling)
         }
@@ -309,32 +361,33 @@ class ValiderOmregningServiceTest {
         private fun innvilge(vararg perioder: VedtaksperiodeDto) =
             InnvilgelseOvergangsstønad(null, null, perioder.toList())
 
-        private fun manuellGOmregning() = saksbehandling(årsak = BehandlingÅrsak.G_OMREGNING).copy(opprettetAv = "saksbehandler")
+        private fun manuellGOmregning() =
+            saksbehandling(årsak = BehandlingÅrsak.G_OMREGNING).copy(opprettetAv = "saksbehandler")
     }
 
     private fun mockNyTilkjentYtelse(saksbehandling: Saksbehandling, medRiktigBeløp: Boolean = true) {
         every { tilkjentYtelseRepository.findByBehandlingId(saksbehandling.id) } returns
-            lagTilkjentYtelse(
-                andelerTilkjentYtelse = listOf(
-                    lagAndelTilkjentYtelse(
-                        fraOgMed = LocalDate.of(2022, 4, 1),
-                        tilOgMed = LocalDate.of(2022, 4, 30),
-                        samordningsfradrag = 5000,
-                        beløp = 14950,
+                lagTilkjentYtelse(
+                    andelerTilkjentYtelse = listOf(
+                        lagAndelTilkjentYtelse(
+                            fraOgMed = LocalDate.of(2022, 4, 1),
+                            tilOgMed = LocalDate.of(2022, 4, 30),
+                            samordningsfradrag = 5000,
+                            beløp = 14950,
+                        ),
+                        lagAndelTilkjentYtelse(
+                            fraOgMed = LocalDate.of(2022, 5, 1),
+                            tilOgMed = LocalDate.of(2022, 8, 30),
+                            samordningsfradrag = 5000,
+                            beløp = if (medRiktigBeløp) 15902 else 0,
+                        ),
                     ),
-                    lagAndelTilkjentYtelse(
-                        fraOgMed = LocalDate.of(2022, 5, 1),
-                        tilOgMed = LocalDate.of(2022, 8, 30),
-                        samordningsfradrag = 5000,
-                        beløp = if (medRiktigBeløp) 15902 else 0,
-                    ),
-                ),
-            )
+                )
     }
 
     private fun mockVedtakOgForrigeTilkjentYtelse(saksbehandling: Saksbehandling) {
         every { tilkjentYtelseRepository.findByBehandlingId(saksbehandling.forrigeBehandlingId!!) } returns
-            lagTilkjentYtelse(emptyList(), grunnbeløpsmåned = YearMonth.of(2021, 5))
+                lagTilkjentYtelse(emptyList(), grunnbeløpsmåned = YearMonth.of(2021, 5))
         every { vedtakService.hentVedtak(saksbehandling.id) } returns vedtak(saksbehandling.id)
     }
 }
